refactor: usePreviousRouteName hook
